Extract ProgressSummaryCompleted to own module
